+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Edge
 
-* No new features yet.
+* The `servers` key in the YAML configuration file is now a hash instead of an array.
 
 ## v0.3.0 - October 25, 2013